moved timer components to timer folder
