+++ app/components/timer/progressBar.js
