+++ app/components/timer/timeControls.js
